[Analysis, C++] InputManager written in C++
The InputManager files (not a class at the moment) were written in C++
in order to encrease performance. At this point there are two main
problems:
o) since it uses variadic templates, it strictly requires PyROOT
experimental in order to make the injection happen properly; at the
moment of writing, it's not possible to compile ROOT with PyROOT
experimental due to some git issues that prevent me to test the Python
side of the main properly;
o) in this phase, the C++ code is not split into .h and .cxx (everything
is in .h) due to some (more) technical problems: by running PyROOT
(current) in this machine (after sourcing cvmfs) I get a SEG_FAULT if
the declaration is separated from the definition; I'm not able to
reproduce this situation on my laptop, that's why a JIRA ticket wasn't
opened. Needs to be investigated.