[Analysis, Python] CutVec derived from class list
Instead of having cutvec as an attribute of CutVec (horrible), CutVec is
now a class derived from list, so it has the behavior of a list.
However, the best idea would be for it to be derived from
ROOT.VecOps.RVec, in order to have zero-copy operations when
manipulated.
This probably needs to be done entirely on C++ side, because:
o) a Python class (Cut) can't be passed as a template tye to RVec;
o) RVec does not implement a virtual destructor, and with new Cppyy it
is required to have one in order to derive a Python class from a C++
class.