[Analysis, Python] 'Channel' class introduced
Channel is derived from CutScheduler, and introduces some more
attributes such as name (em, mt, tt, ecc.), era (2016-17-18) and model
(temporary, SM MSSM).
With the function 'create_channels', the channels are created based on
the ones passed as parameter through the command line.

N.B.: this is temporary and for now only "em" cuts are written;
the best idea would probably be to have the chosen cuts NOT hardcoded,
but passed in a JSON file.
This needs to be discussed.